p2/api/requirements.txt
django-filter
django-rest-framework
djangorestframework-guardian
djangorestframework-jwt
drf-yasg[validation]
